+++ Changelog.md
@@ -19,6 +19,7 @@
 - Cache playwright's chromium installation on GitHub Actions (#7372)
 - Fix broken link to the Vagrant installation guide in `README.md` (#7349)
 - Fix `extra_hosts` configuration in `compose.yaml` (#7375)
+- Convert front-end tests from enzyme to react testing library; add `@testing-library/user-event` (#7379)
 
 ## [v2.6.1]